% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.add_bmi.R
\name{add_bmi}
\alias{add_bmi}
\title{Add BMI to Data Frame}
\usage{
add_bmi(
  data,
  age,
  normed_bmi,
  data_source = "cdc",
  type = "bmiz",
  sex = 2,
  dob = NULL,
  date_assessed = NULL,
  age_unit = NULL
)
}
\arguments{
\item{data}{A data frame containing the necessary columns.}

\item{age}{A column in \code{data} representing the age of individuals.}

\item{normed_bmi}{A column in \code{data} representing the normalized BMI (either BMI percentile or BMI z-score).}

\item{data_source}{A string specifying the data source, either 'cdc' or 'wgsr'. Default is 'cdc'.}

\item{type}{A string specifying the type of normalized BMI, either 'bmiz' for BMI z-score or 'pct' for BMI percentile. Default is 'bmiz'.}

\item{sex}{A column in \code{data} representing the sex of individuals, where 1 = Male and 2 = Female. Default is 2 (Female).}

\item{dob}{(Optional) A column in \code{data} representing the date of birth.}

\item{date_assessed}{(Optional) A column in \code{data} representing the date of assessment.}

\item{age_unit}{(Optional) A string specifying the unit of the age column.}
}
\value{
A data frame with an added BMI column.
}
\description{
This function adds a BMI column to a data frame based on age, sex, and normalized BMI (BMI percentile or BMI z-score).
}
