% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.bmi_forecast.R
\name{add_eBMI_to_df}
\alias{add_eBMI_to_df}
\title{Add eBMI to Data Frame}
\usage{
add_eBMI_to_df(data, adult_height = adult_height_in)
}
\arguments{
\item{data}{A data frame containing BMI data.}

\item{adult_height}{Column name for adult height.}
}
\value{
A data frame with added eBMI and weight columns.
}
\description{
Add estimated BMI (eBMI) and corresponding weights to the data frame.
}
