% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.age_math.R
\name{age_in_days}
\alias{age_in_days}
\title{Convert age from various units to days.}
\usage{
age_in_days(age = NULL, age_unit = NULL, dob = NULL, date_assessed = NULL)
}
\arguments{
\item{age}{Numeric value representing the age.}

\item{age_unit}{(Optional) A character string specifying the unit of the input age ('years', 'months', 'days').}

\item{dob}{(Optional) The date of birth as a character string in the format 'YYYY-MM-DD'.}

\item{date_assessed}{(Optional) The date of assessment as a character string in the format 'YYYY-MM-DD'.}
}
\value{
Numeric value representing the age in days.
}
\description{
This function takes an age input that can be in years, months, days, or as a combination of birthdate and date of assessment.
}
\examples{
age_in_days(10, "years")  # Returns the age in days
age_in_days(120, "months")  # Returns the age in days

}
