% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.age_math.R
\name{age_unit}
\alias{age_unit}
\title{Determine the age unit (days, weeks, months, or years) based on the input age.}
\usage{
age_unit(age)
}
\arguments{
\item{age}{Numeric value representing the age.}
}
\value{
A character string indicating the age unit (days, weeks, months, or years).
}
\description{
This function assumes that the input age is between 5 and 21 years.
}
\examples{
age_unit(365)  # Returns "days"
age_unit(10*365)  # Returns "years"

}
