% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.sex_coding.R
\name{align_sex_coding}
\alias{align_sex_coding}
\title{Align Sex Coding}
\usage{
align_sex_coding(sex)
}
\arguments{
\item{sex}{A vector of sex values coded as 'Male/Female', 'M/F', '0(Male)/1(Female)', or '1(Male)/2(Female)'.}
}
\value{
A list containing the detected coding system and the aligned sex vector with values 1 for male and 2 for female.
}
\description{
This function detects how the sex variable is coded ('Male/Female', 'M/F', or '0(Male)/1(Female)') and converts them to 1/2.
}
\examples{
sex_vector <- c("Male", "Female", "Male")
align_sex_coding(sex_vector)

}
