% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.sex_coding.R
\name{align_sex_coding_in_df}
\alias{align_sex_coding_in_df}
\title{Align Sex Coding in Data Frame}
\usage{
align_sex_coding_in_df(data, sex_column = "sex")
}
\arguments{
\item{data}{A data frame containing a column with sex values.}

\item{sex_column}{The name of the column containing the sex values. Defaults to 'sex'.}
}
\value{
The data frame with the sex column values aligned to 1 for male and 2 for female.
}
\description{
This function aligns the sex coding in a data frame by detecting how the sex variable is coded ('Male/Female', 'M/F', or '0(Male)/1(Female)') and converting them to 1/2.
}
\examples{
data <- data.frame(sex = c("Male", "Female", "Male"), age = c(25, 30, 35))
align_sex_coding_in_df(data)

}
