% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.bmi_forecast.R
\name{apply_bmi_lookup}
\alias{apply_bmi_lookup}
\title{Apply BMI Lookup}
\usage{
apply_bmi_lookup(
  data,
  data_point_col,
  age_col = agemos,
  data_source = "cdc",
  type = "bmiz"
)
}
\arguments{
\item{data}{A data frame containing BMI data.}

\item{data_point_col}{Column name for data points.}

\item{age_col}{Column name for age.}

\item{data_source}{Data source for BMI lookup (default: 'cdc').}

\item{type}{Type of BMI data (default: 'bmiz').}
}
\value{
A vector of BMI values.
}
\description{
Apply the BMI lookup function to a data frame.
}
