% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.bmiz_to_bmi.R
\name{bmi_lookup}
\alias{bmi_lookup}
\title{Lookup BMI}
\usage{
bmi_lookup(
  data_point,
  type = "bmiz",
  data_source = "cdc",
  sex = 2,
  age,
  age_unit = NULL,
  dob = NULL,
  date_assessed = NULL
)
}
\arguments{
\item{data_point}{Data point (BMI Z-score or percentile).}

\item{type}{Type of data point ('bmiz' or 'pct').}

\item{data_source}{Data source ('cdc' or 'wgsr').}

\item{sex}{Sex of the individual (1 for male, 2 for female).}

\item{age}{Age of the individual.}

\item{age_unit}{Unit of age ('months', 'years', etc.).}

\item{dob}{Date of birth (optional).}

\item{date_assessed}{Date of assessment (optional).}
}
\value{
BMI value corresponding to the given data point.
}
\description{
Lookup BMI
}
