% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.bmiz_to_bmi.R
\name{bmi_to_bmiz}
\alias{bmi_to_bmiz}
\title{Convert BMI to Z-Score}
\usage{
bmi_to_bmiz(l, m, s, bmi, sigma = NULL, p95 = NULL)
}
\arguments{
\item{l}{L parameter from the LMS table.}

\item{m}{M parameter from the LMS table.}

\item{s}{S parameter from the LMS table.}

\item{bmi}{BMI value.}

\item{sigma}{Standard deviation (optional).}

\item{p95}{95th percentile BMI value (optional).}
}
\value{
Z-score corresponding to the given BMI.
}
\description{
Convert BMI to Z-Score
}
