% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.bmiz_to_bmi.R
\name{bmiz_lookup}
\alias{bmiz_lookup}
\title{Lookup BMI Z-Score}
\usage{
bmiz_lookup(
  bmi,
  sex = 2,
  age,
  data_source = "cdc",
  age_unit = NULL,
  dob = NULL,
  date_assessed = NULL
)
}
\arguments{
\item{bmi}{BMI value.}

\item{sex}{Sex of the individual (1 for male, 2 for female).}

\item{age}{Age of the individual.}

\item{data_source}{Data source ('cdc' or 'wgsr').}

\item{age_unit}{Unit of age ('months', 'years', etc.).}

\item{dob}{Date of birth (optional).}

\item{date_assessed}{Date of assessment (optional).}
}
\value{
Z-score corresponding to the given BMI.
}
\description{
Lookup BMI Z-Score
}
