% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.bmiz_to_bmi.R
\name{bmiz_to_bmi}
\alias{bmiz_to_bmi}
\title{Convert Z-Score to BMI}
\usage{
bmiz_to_bmi(l, m, s, bmiz, sigma = NULL, p95 = NULL)
}
\arguments{
\item{l}{L parameter from the LMS table.}

\item{m}{M parameter from the LMS table.}

\item{s}{S parameter from the LMS table.}

\item{bmiz}{BMI Z-score.}

\item{sigma}{Standard deviation (optional).}

\item{p95}{95th percentile BMI value (optional).}
}
\value{
BMI corresponding to the given Z-score.
}
\description{
Convert Z-Score to BMI
}
