% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_bmi.r
\name{calculate_bmi}
\alias{calculate_bmi}
\title{Function that calculates BMI (Body Mass Index) based on weight and height inputs.}
\usage{
calculate_bmi(weight, weight_unit = "kg", height, height_unit = "cm")
}
\arguments{
\item{weight}{Weight value in either kilograms (kg) or pounds (lbs).}

\item{weight_unit}{Unit of weight measurement ('kg' or 'lbs').}

\item{height}{Height value in either centimeters (cm), meters (m), inches (in), or feet (ft).}

\item{height_unit}{Unit of height measurement ('cm', 'm', 'in', or 'ft').}
}
\value{
BMI calculated based on the provided weight and height.
}
\description{
Function that calculates BMI (Body Mass Index) based on weight and height inputs.
}
