% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdcData.R
\docType{data}
\name{cdcData}
\alias{cdcData}
\title{CDC Growth Data}
\format{
An object of class \code{data.frame} with 438 rows and 35 columns.
}
\usage{
data(cdcData)
}
\description{
A dataset containing CDC growth data used for calculating BMI and other growth metrics.
}
\keyword{datasets}
