% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.bmi_forecast.R
\name{clean_and_process}
\alias{clean_and_process}
\title{Clean and Process Data}
\usage{
clean_and_process(
  data,
  id_col_name = NULL,
  age_col_name = NULL,
  dob_col_name = NULL,
  date_assessed_col_name = NULL,
  age_unit = NULL,
  sex_col_name = NULL,
  ht_col_name = NULL,
  ht_unit = "cm",
  wt_col_name = NULL,
  wt_unit = "kg",
  bmi_col_name = NULL,
  bmiz_col_name = NULL,
  pct_col_name = NULL,
  data_source = "cdc",
  adult_height_col_name = NULL,
  lower_margin = 0.5,
  upper_margin = 0.5,
  central_value = "mean"
)
}
\arguments{
\item{data}{A data frame containing BMI data.}
}
\value{
A full BMI data frame with processed forecasts and cutoffs.
}
\description{
Clean and Process Data
}
