% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.clean_raw_data.R
\name{clean_data}
\alias{clean_data}
\title{Clean BMI data
This function intakes raw data and creates columns for id, agemos, sex, adult height, BMI, and BMIZ. It handles various formats for sex and height/weight units, calculates age in months, and manages missing columns.}
\usage{
clean_data(
  data,
  id_col_name = NULL,
  age_col_name = NULL,
  dob_col_name = NULL,
  date_assessed_col_name = NULL,
  age_unit = NULL,
  sex_col_name = NULL,
  ht_col_name = NULL,
  ht_unit = "cm",
  wt_col_name = NULL,
  wt_unit = "kg",
  bmi_col_name = NULL,
  bmiz_col_name = NULL,
  pct_col_name = NULL,
  data_source = "cdc",
  adult_height_col_name = NULL,
  age_adult_ht_col_name = NULL,
  ed_aoo_col_name = NULL
)
}
\arguments{
\item{data}{A data frame containing the raw data.}

\item{id_col_name}{The name of the column containing unique identifiers. Defaults to NULL.}

\item{age_col_name}{The name of the column containing age values. Defaults to NULL.}

\item{dob_col_name}{The name of the column containing date of birth. Defaults to NULL.}

\item{date_assessed_col_name}{The name of the column containing the date of assessment. Defaults to NULL.}

\item{age_unit}{The unit of age values ('days', 'months', or 'years'). Defaults to NULL.}

\item{sex_col_name}{The name of the column containing sex values. Defaults to NULL.}

\item{ht_col_name}{The name of the column containing height values. Defaults to NULL.}

\item{ht_unit}{The unit of height values ('cm' or 'inches'). Defaults to 'cm'.}

\item{wt_col_name}{The name of the column containing weight values. Defaults to NULL.}

\item{wt_unit}{The unit of weight values ('kg' or 'lbs'). Defaults to 'kg'.}

\item{bmi_col_name}{The name of the column containing BMI values. Defaults to NULL.}

\item{bmiz_col_name}{The name of the column containing BMIZ values. Defaults to NULL.}

\item{pct_col_name}{The name of the column containing percentile values. Defaults to NULL.}

\item{data_source}{The source of the data ('cdc' or other). Defaults to 'cdc'.}

\item{adult_height_col_name}{The name of the column containing adult height values. Defaults to NULL.}
}
\value{
A data frame with columns for id, agemos, sex, adult_height, bmi, and bmiz.
}
\description{
Clean BMI data
This function intakes raw data and creates columns for id, agemos, sex, adult height, BMI, and BMIZ. It handles various formats for sex and height/weight units, calculates age in months, and manages missing columns.
}
\examples{
raw_data <- data.frame(id = 1:5, sex = c("M", "F", "M", "F", "M"), age = c(16, 17, 18, 19, 20), height = c(160, 165, 170, 175, 180), weight = c(60, 65, 70, 75, 80))
clean_data(raw_data, id_col_name = "id", age_col_name = "age", sex_col_name = "sex", ht_col_name = "height", wt_col_name = "weight")

}
