% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.make_cutoff_data.R
\name{cutoff_bmi_by_age_data}
\alias{cutoff_bmi_by_age_data}
\title{Cutoff BMI}
\usage{
cutoff_bmi_by_age_data(
  age = agemos,
  data_source = "cdc",
  sex = 2,
  adult_height = NA
)
}
\arguments{
\item{age}{'age in months'}

\item{data_source}{'cdc' or 'who'}

\item{sex}{'1 = Male, 2 = Female'}

\item{adult_height}{'adult height in inches'}
}
\value{
bmi scores based on bmiz and age
}
\description{
Makes Cutoff BMIs
}
