% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.bmi_forecast.R
\name{cutoffs_by_participant}
\alias{cutoffs_by_participant}
\title{Generate BMI Cutoffs by Participant}
\usage{
cutoffs_by_participant(data)
}
\arguments{
\item{data}{A data frame containing participant information.}
}
\value{
A data frame with BMI cutoff data for each participant.
}
\description{
Generate BMI cutoff data for each participant in the data frame.
}
