% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\docType{data}
\name{demo}
\alias{demo}
\title{Demo Data}
\format{
A data frame with the following columns:
\describe{
\item{participant}{Unique identifier for each individual}
\item{age}{Age in years}
\item{sex}{Sex (M = male, F = female)}
\item{weight}{Weight in kilograms}
\item{height}{Height in centimeters}
\item{adult_height_in}{Adult height in inches}
\item{ed_aoo}{Age of onset of eating disorder}
}
}
\usage{
data(demo)
}
\description{
A sample dataset used for demonstrating package functions.
}
\keyword{datasets}
