% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.bmi_forecast.R
\name{forecast_bmi}
\alias{forecast_bmi}
\title{Make Full BMI Data Frame}
\usage{
forecast_bmi(
  data,
  ci,
  adult_height = "adult_height_in",
  id = "id",
  lower_margin = 0.5,
  upper_margin = 0.5,
  central_value = "mean"
)
}
\arguments{
\item{data}{A cleaned data frame containing BMI data.}

\item{ci}{The confidence interval level.}

\item{adult_height}{Column name for adult height (default: 'adult_height').}

\item{id}{Column name for IDs (default: 'id').}

\item{lower_margin}{The lower margin for the user-defined forecast.}

\item{upper_margin}{The upper margin for the user-defined forecast.}

\item{central_value}{The central value for the user-defined forecast.}
}
\value{
A full BMI data frame with processed forecasts and cutoffs.
}
\description{
Create a full BMI data frame by processing forecasts, generating cutoffs, and adding eBMI values.
}
