% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.calculate_bmi.r
\name{height_in_cm}
\alias{height_in_cm}
\title{Function that converts height from various units (inches, meters, feet) to centimeters (cm).}
\usage{
height_in_cm(height, height_unit = "cm")
}
\arguments{
\item{height}{Height value in either inches (in), meters (m), or feet (ft).}

\item{height_unit}{Unit of height measurement ('in', 'm', or 'ft').}
}
\value{
Height value converted to centimeters (cm).
}
\description{
Function that converts height from various units (inches, meters, feet) to centimeters (cm).
}
