% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.bmiz_forecast.R
\name{make_bmiz_forecast}
\alias{make_bmiz_forecast}
\title{Create and Forecast BMIz Data}
\usage{
make_bmiz_forecast(
  data,
  ci = 95,
  lower_margin = NULL,
  upper_margin = NULL,
  central_value = c("mean", "max", "most_recent", "mean+most_recent")
)
}
\arguments{
\item{data}{A data frame containing the input data.}

\item{ci}{Confidence interval for the forecast.}

\item{lower_margin}{Lower margin for the forecast.}

\item{upper_margin}{Upper margin for the forecast.}

\item{central_value}{Central value for the forecast. Options are "mean", "max", "most_recent", "mean+most_recent".}
}
\value{
A forecast object for each id.
}
\description{
Create and Forecast BMIz Data
}
