% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.bmiz_forecast.R
\name{make_bmiz_tsibble}
\alias{make_bmiz_tsibble}
\title{Create tsibble for BMIz Data}
\usage{
make_bmiz_tsibble(data)
}
\arguments{
\item{cleaned_data}{A data frame containing the input data - cleaned per clean_data function}

\item{dob}{Column name for date of birth (optional).}

\item{date_assessed}{Column name for date of assessment (optional).}

\item{age_unit}{Unit of age (optional).}
}
\value{
A tsibble with the BMIz data.
}
\description{
Create tsibble for BMIz Data
}
