% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9.Percentile_Restore_Plot.R
\name{Pct_Restore_Plot}
\alias{Pct_Restore_Plot}
\title{BMI Percentile Restoration Planning Plot}
\usage{
Pct_Restore_Plot(clean_data, forecast_data)
}
\arguments{
\item{clean_data}{a cleaned dataframe - cleaned by the tx_plot_clean function}

\item{forecast_data}{a forecasted dataframe for 1 participant}

\item{slope_per_week}{the slope of weight gain expected in lbs per week}
}
\value{
a treatment plan plot
}
\description{
BMI Percentile Restoration Planning Plot
}
