% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.bmiz_to_bmi.R
\name{pct_to_bmiz}
\alias{pct_to_bmiz}
\title{Convert Percentile to Z-Score}
\usage{
pct_to_bmiz(pct)
}
\arguments{
\item{pct}{Percentile value (0-100).}
}
\value{
Z-score corresponding to the given percentile.
}
\description{
Convert Percentile to Z-Score
}
