% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_BMI.R
\name{Plot_eBMI}
\alias{Plot_eBMI}
\title{Plot BMI}
\usage{
Plot_eBMI(data, age, bmi, sex)
}
\arguments{
\item{data}{'data'}

\item{age}{'age in months'}

\item{sex}{'1 = Male, 2 = Female'}

\item{age_days}{'age in days column'}

\item{adult_height}{'adult height in cm'}
}
\value{
bmi scores based on bmiz and age
}
\description{
Plot BMI
}
