% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.Plot_eBMI.R
\name{plot_weight}
\alias{plot_weight}
\title{Plot Weight Forecast}
\usage{
plot_weight(
  clean_data,
  forecast_data,
  px,
  agemos_ed_onset = NA,
  a_height = NULL
)
}
\arguments{
\item{forecast_data}{A data frame containing the forecast data.}

\item{px}{The participant ID to filter the data.}

\item{a_height}{Optional parameter to specify the adult height if it is not available in the data.}

\item{model}{The model type used for forecasting.}

\item{ci}{The confidence interval level.}
}
\value{
A ggplot object displaying the weight forecast.
}
\description{
This function plots the weight forecast data for a given participant, model, and confidence interval.
}
\examples{
# Assuming forecast_data is your dataframe and embarktools is loaded
# plot_weight(forecast_data, "ARIMA", "95\%", 1, a_height = 65)
}
