% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.bmi_forecast.R
\name{process_bmiz_forecast}
\alias{process_bmiz_forecast}
\title{Process BMIZ Forecast Data Frame}
\usage{
process_bmiz_forecast(
  df,
  model,
  ci,
  lower_margin = NULL,
  upper_margin = NULL,
  central_value = NULL
)
}
\arguments{
\item{df}{A BMIZ forecast data frame.}

\item{model}{The model type used for forecasting.}

\item{ci}{The confidence interval level.}

\item{lower_margin}{The lower margin for the user-defined forecast.}

\item{upper_margin}{The upper margin for the user-defined forecast.}

\item{central_value}{The central value for the user-defined forecast.}
}
\value{
A processed data frame with formatted intervals.
}
\description{
Convert \code{hilo} objects to character strings and format intervals.
}
