% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.make_cutoff_data.R
\name{solve_for_weight}
\alias{solve_for_weight}
\title{solve for weight}
\usage{
solve_for_weight(bmi, height)
}
\arguments{
\item{bmi}{'bmi'}

\item{height}{'height in inches'}
}
\value{
weight
}
\description{
Using BMI and height, solve for weight
}
