% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.make_cutoff_data.R
\name{solve_for_weight_metric}
\alias{solve_for_weight_metric}
\title{solve for weight}
\usage{
solve_for_weight_metric(bmi, height)
}
\arguments{
\item{bmi}{'bmi'}

\item{height}{'height in cm'}
}
\value{
weight
}
\description{
Using BMI and height, solve for weight
}
