% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.Clean_WtRestore.R
\name{tx_plot_clean}
\alias{tx_plot_clean}
\title{Clean Treatment Data}
\usage{
tx_plot_clean(
  filtered_data,
  age_col_name = NULL,
  date_assessed_col_name = NULL,
  ht_col_name = NULL,
  wt_col_name = NULL,
  age_unit = NULL,
  ht_unit = "cm",
  wt_unit = "kg",
  bmi_col_name = NULL,
  bmiz_col_name = NULL,
  pct_col_name = NULL,
  data_source = "cdc",
  dob = NULL,
  sex = NULL,
  current_ht = NULL,
  age_current_ht = NULL,
  adult_ht = NULL,
  age_adult_ht = NULL,
  ed_aoo = NULL,
  tx_start_date = NULL,
  intake_wt = NULL
)
}
\arguments{
\item{filtered_data}{A data frame containing the filtered data for a single participant.}

\item{age_col_name}{The name of the column containing the age of the participant}

\item{date_assessed_col_name}{The name of the column containing the date the data was assessed}

\item{ht_col_name}{The name of the column containing the height of the participant}

\item{wt_col_name}{The name of the column containing the weight of the participant}

\item{age_unit}{The unit of the age column (e.g., years, months, days)}

\item{ht_unit}{The unit of the height column (e.g., cm, in)}

\item{wt_unit}{The unit of the weight column (e.g., kg, lb)}

\item{bmi_col_name}{The name of the column containing the BMI of the participant}

\item{bmiz_col_name}{The name of the column containing the BMI z-score of the participant}

\item{pct_col_name}{The name of the column containing the percentile of the participant}

\item{data_source}{The source of the data (e.g., cdc, who)}

\item{dob}{The date of birth of the participant}

\item{sex}{Sex}

\item{current_ht}{The current height of the participant}

\item{age_current_ht}{The age at which the current height was measured}

\item{adult_ht}{The adult height of the participant}

\item{age_adult_ht}{The age at which the adult height was measured}

\item{ed_aoo}{The age of onset of the eating disorder}

\item{tx_start_date}{The start date of the treatment}
}
\value{
A cleaned and processed data frame with the age in years, expected
BMI, and expected weight, formatted appropriately.
}
\description{
Clean Treatment Data
}
