% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgsrData.R
\docType{data}
\name{wgsrData}
\alias{wgsrData}
\title{WHO Growth Standards Data}
\format{
An object of class \code{data.frame} with 28992 rows and 6 columns.
}
\usage{
data(wgsrData)
}
\description{
A dataset containing WHO growth standards used for calculating BMI and other growth metrics.
}
\keyword{datasets}
