% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.Wt_Restore_Plot.R
\name{Wt_Restore_Plot}
\alias{Wt_Restore_Plot}
\title{Weight Restoration Planning Plot}
\usage{
Wt_Restore_Plot(clean_data, forecast_data, slope_per_week = 0.5)
}
\arguments{
\item{clean_data}{a cleaned dataframe - cleaned by the tx_plot_clean function}

\item{forecast_data}{a forecasted dataframe for 1 participant}

\item{slope_per_week}{the slope of weight gain expected in lbs per week}
}
\value{
a treatment plan plot
}
\description{
Weight Restoration Planning Plot
}
