% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bmi.R
\name{add_bmi}
\alias{add_bmi}
\title{Add BMI to data frame
@param data Data frame
@param age Age column
@param normed_bmi Normalized BMI column (either bmi percentile or bmiz score)
@param data_source Data source ('cdc' or 'wgsr') - Default to 'cdc'
@param type 'bmiz' or 'pct' - Default to 'bmiz'
@param sex sex - 1 = Male; 2 = Female - Default to Female
@param dob Date of birth column (optional)
@param date_assessed Date assessed column (optional)
@param age_unit Unit of age column (optional)
@import dplyr
@return data frame with BMI column added
@export}
\usage{
add_bmi(
  data,
  age,
  normed_bmi,
  data_source = "cdc",
  type = "bmiz",
  sex = 2,
  dob = NULL,
  date_assessed = NULL,
  age_unit = NULL
)
}
\description{
Add BMI to data frame
@param data Data frame
@param age Age column
@param normed_bmi Normalized BMI column (either bmi percentile or bmiz score)
@param data_source Data source ('cdc' or 'wgsr') - Default to 'cdc'
@param type 'bmiz' or 'pct' - Default to 'bmiz'
@param sex sex - 1 = Male; 2 = Female - Default to Female
@param dob Date of birth column (optional)
@param date_assessed Date assessed column (optional)
@param age_unit Unit of age column (optional)
@import dplyr
@return data frame with BMI column added
@export
}
