% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.age_math.R
\name{age_in_months}
\alias{age_in_months}
\title{Convert age in days to age in months (rounded).}
\usage{
age_in_months(age, age_unit = NULL, dob = NULL, date_assessed = NULL)
}
\arguments{
\item{age_days}{Numeric value representing the age in days.}
}
\value{
Numeric value representing the age in months (rounded to the nearest month).
}
\description{
Convert age in days to age in months (rounded).
}
\examples{
age_in_months(365)  # Returns 12
age_in_months(730)  # Returns 24

}
