% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bmiz.R
\name{align_sex_coding}
\alias{align_sex_coding}
\title{Detect how sex variable is coded and convert it to 1/2.}
\usage{
align_sex_coding(sex)
}
\arguments{
\item{sex}{A vector containing the sex variable with different codings.}
}
\value{
A list with 'coding' indicating the coding type and 'sex' with numeric values.
}
\description{
This function detects the coding of the sex variable and converts it to numeric values (1 for Male, 2 for Female).
}
\examples{
align_sex_coding(c("Male", "Female"))  # Returns list(coding = "Male/Female", sex = c(1, 2))
}
