% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.bmiz_to_bmi.R
\name{bmi_lookup}
\alias{bmi_lookup}
\title{BMI Lookup}
\usage{
bmi_lookup(
  data_point,
  type = "bmiz",
  data_source = "cdc",
  sex = 2,
  age,
  age_unit = NULL,
  dob = NULL,
  date_assessed = NULL
)
}
\arguments{
\item{data_point}{'bmiz score or bmi percentile'}

\item{type}{'bmiz' or 'pct'}

\item{data_source}{Data source ('cdc' or 'wgsr')}

\item{sex}{'1 = Male, 2 = Female'}

\item{age}{'age'}

\item{age_unit}{'unit of age - optional'}

\item{dob}{'date of birth'}

\item{date_assessed}{'date of assessment}
}
\value{
bmi scores based on bmiz and age
}
\description{
BMI Lookup
}
