% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.bmiz_to_bmi.R
\name{bmi_to_bmiz}
\alias{bmi_to_bmiz}
\title{Converts bmi to bmiz scores based on either cdc or WGSR lookup data}
\usage{
bmi_to_bmiz(l, m, s, bmi)
}
\arguments{
\item{l}{matching l parameter from CDC or WGSR data}

\item{m}{from CDC or WGSR data}

\item{s}{from CDC or WGSR data}

\item{bmi}{bmi score from original data}
}
\value{
a bmiz
}
\description{
Converts bmi to bmiz scores based on either cdc or WGSR lookup data
}
