% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.bmiz_to_bmi.R
\name{bmiz_lookup}
\alias{bmiz_lookup}
\title{BMIz Lookup}
\usage{
bmiz_lookup(
  bmi,
  sex = 2,
  age,
  data_source = "cdc",
  age_unit = NULL,
  dob = NULL,
  date_assessed = NULL
)
}
\arguments{
\item{bmi}{'bmi score'}

\item{sex}{'1 = Male, 2 = Female'}

\item{age}{'age'}

\item{data_source}{Data source ('cdc' or 'wgsr')}

\item{age_unit}{'unit of age - optional'}

\item{dob}{'date of birth'}

\item{date_assessed}{'date of assessment'}
}
\value{
bmiz scores based on bmi and age
}
\description{
BMIz Lookup
}
