% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.bmiz_to_bmi.R
\name{bmiz_to_bmi}
\alias{bmiz_to_bmi}
\title{Converts bmiz scores to bmi based on either cdc or WGSR lookup data}
\usage{
bmiz_to_bmi(l, m, s, bmiz)
}
\arguments{
\item{l}{matching l parameter from CDC or WGSR data}

\item{m}{from CDC or WGSR data}

\item{s}{from CDC or WGSR data}

\item{bmiz}{bmiz score from original data}
}
\value{
a bmi
}
\description{
Converts bmiz scores to bmi based on either cdc or WGSR lookup data
}
