% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.bmi_forecast.R
\name{cull_demos}
\alias{cull_demos}
\title{Cull Demographics Data}
\usage{
cull_demos(data)
}
\arguments{
\item{data}{A cleaned data frame containing demographic information.}
}
\value{
A data frame with aligned demographics data.
}
\description{
Align participant, sex coding, and add adult height and sex columns, if available.
}
