% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmiz_forecast.R
\name{fit_and_forecast_bmiz}
\alias{fit_and_forecast_bmiz}
\title{Fit Models and Generate Forecasts for BMIz Data}
\usage{
fit_and_forecast_bmiz(
  ts_data,
  lower_margin = 0.5,
  upper_margin = 0.5,
  central_value = "mean"
)
}
\arguments{
\item{ts_data}{A tsibble containing the BMIz data.}
}
\value{
A forecast object for each participant.
}
\description{
Fit Models and Generate Forecasts for BMIz Data
}
