% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.bmiz_forecast.R
\name{fit_and_forecast_bmiz}
\alias{fit_and_forecast_bmiz}
\title{Fit Models and Generate Forecasts for BMIz Data}
\usage{
fit_and_forecast_bmiz(
  ts_data,
  lower_margin = lower_margin,
  upper_margin = upper_margin,
  central_value = central_value
)
}
\arguments{
\item{ts_data}{A tsibble containing the BMIz data.}
}
\value{
A forecast object for each id.
}
\description{
Fit Models and Generate Forecasts for BMIz Data
}
