% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmiz_forecast.R
\name{make_bmiz_forecast}
\alias{make_bmiz_forecast}
\title{Create and Forecast BMIz Data}
\usage{
make_bmiz_forecast(
  data,
  age = age,
  bmiz = bmiz,
  participant = NULL,
  dob = NULL,
  date_assessed = NULL,
  age_unit = NULL,
  lower_margin = 0.5,
  upper_margin = 0.5,
  central_value = "mean"
)
}
\arguments{
\item{data}{A data frame containing the input data.}

\item{age}{Column name for age in the data frame.}

\item{bmiz}{Column name for BMIz in the data frame.}

\item{participant}{Column name for participant identifier.}

\item{dob}{Column name for date of birth (optional).}

\item{date_assessed}{Column name for date of assessment (optional).}

\item{age_unit}{Unit of age (optional).}
}
\value{
A forecast object for each participant.
}
\description{
Create and Forecast BMIz Data
}
