% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.bmiz_forecast.R
\name{make_bmiz_forecast}
\alias{make_bmiz_forecast}
\title{Create and Forecast BMIz Data}
\usage{
make_bmiz_forecast(
  data,
  lower_margin = lower_margin,
  upper_margin = upper_margin,
  central_value = central_value
)
}
\arguments{
\item{data}{A data frame containing the input data.}

\item{age}{Column name for age in the data frame.}

\item{bmiz}{Column name for BMIz in the data frame.}

\item{id}{Column name for id identifier.}

\item{dob}{Column name for date of birth (optional).}

\item{date_assessed}{Column name for date of assessment (optional).}

\item{age_unit}{Unit of age (optional).}
}
\value{
A forecast object for each id.
}
\description{
Create and Forecast BMIz Data
}
