% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.bmiz_forecast.R
\name{make_bmiz_tsibble}
\alias{make_bmiz_tsibble}
\title{Create tsibble for BMIz Data}
\usage{
make_bmiz_tsibble(
  data,
  age = age,
  bmiz = bmiz,
  id = NULL,
  dob = NULL,
  date_assessed = NULL,
  age_unit = NULL
)
}
\arguments{
\item{data}{A data frame containing the input data.}

\item{age}{Column name for age in the data frame.}

\item{bmiz}{Column name for BMIz in the data frame.}

\item{id}{Column name for id identifier.}

\item{dob}{Column name for date of birth (optional).}

\item{date_assessed}{Column name for date of assessment (optional).}

\item{age_unit}{Unit of age (optional).}
}
\value{
A tsibble with the BMIz data.
}
\description{
Create tsibble for BMIz Data
}
