% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.bmiz_forecast.R
\name{make_bmiz_tsibble}
\alias{make_bmiz_tsibble}
\title{Create tsibble for BMIz Data}
\usage{
make_bmiz_tsibble(data)
}
\arguments{
\item{data}{A data frame containing the input data - cleaned per clean_data function}
}
\value{
A tsibble with the BMIz data.
}
\description{
Create tsibble for BMIz Data
}
