% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.Plot_eBMI.R
\name{plot_eBMI}
\alias{plot_eBMI}
\title{plot_eBMI}
\usage{
plot_eBMI(clean_data, forecast_data, px, agemos_ed_onset = NA)
}
\arguments{
\item{clean_data}{'clean data'}

\item{forecast_data}{'forecast data'}

\item{px}{'participant id'}

\item{agemos_ed_onset}{'age of ED onset'}
}
\value{
plot
}
\description{
plot eBMI
}
