% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.Plot_eBMI.R
\name{plot_eBMI}
\alias{plot_eBMI}
\title{plot_eBMI}
\usage{
plot_eBMI(raw_data, forcast_data, model, id)
}
\arguments{
\item{data}{'data'}

\item{agemos}{'age column'}

\item{sex}{'1 = Male, 2 = Female'}

\item{age}{'age in months'}

\item{adult_height}{'adult height in cm'}
}
\value{
plot
}
\description{
plot eBMI
}
