% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.bmi_forecast.R
\name{process_bmiz_forecast}
\alias{process_bmiz_forecast}
\title{Process BMIZ Forecast Data Frame}
\usage{
process_bmiz_forecast(
  df,
  lower_margin = lower_margin,
  upper_margin = upper_margin,
  central_value = central_value
)
}
\arguments{
\item{df}{A BMIZ forecast data frame.}
}
\value{
A processed data frame with formatted intervals.
}
\description{
Convert \code{hilo} objects to character strings, pivot longer, and format intervals.
}
