% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmiz_forecast.R
\name{user_defined_forecast}
\alias{user_defined_forecast}
\title{Fit Models and Generate Forecasts for BMIz Data}
\usage{
user_defined_forecast(
  data,
  lower_margin = 0.5,
  upper_margin = 0.5,
  central_value = "mean"
)
}
\arguments{
\item{data}{A tsibble containing the BMIz data.}

\item{lower_margin}{Lower margin for the forecast.}

\item{upper_margin}{Upper margin for the forecast.}

\item{central_value}{Central value for the forecast.}
}
\value{
A forecast object for each participant.
}
\description{
Fit Models and Generate Forecasts for BMIz Data
}
