% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.calculate_bmi.r
\name{weight_in_kg}
\alias{weight_in_kg}
\title{Function that converts weight from pounds (lb) to kilograms (kg).}
\usage{
weight_in_kg(weight, weight_units = "kg")
}
\arguments{
\item{weight}{Weight value in pounds (lb) or kilograms (kg) (default).}

\item{weight_units}{Unit of weight measurement ('kg' or 'lbs').}
}
\value{
Weight value converted to kilograms (kg).
}
\description{
Function that converts weight from pounds (lb) to kilograms (kg).
}
